requirements.txt
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
google-api-python-client>=2.100.0
google-auth>=2.23.0